Support more swagger Schema subclasses
I was seeing errors with spec not matching on
io.swagger.v3.oas.models.media.JsonSchema rather than the types
seen in navi like ObjectSchema and StringSchema. It seems like the
only reliable way to tell the "type" of the JsonSchema is to call
.getTypes. This method also works for ObjectSchema etc., so I changed
the spec multimethod to dispatch on the result of .getTypes instead of
on the class.

I merged UUIDSchema with the string case because its .getTypes()
= #{"string"}.

See the Schema class for details
https://github.com/swagger-api/swagger-core/blob/95c8253c5239f6e980acd0b2998d037b35eb8761/modules/swagger-models/src/main/java/io/swagger/v3/oas/models/media/Schema.java