Updated linked-hash-map->clj-map to update-kvs.
Since a LinkedHashMap implements HashMap, a LinkedHashMap "just works"
with Clojure, without any interop needed.  So, re-wrote it using
reduce-kv. Changed the position of the map parameter from last to
first in order to be consistent with update-keys and update-vals.

Refactored other code to use update-kvs.

Some minor comment changes.